+++ .github/workflows/test_openvino.yml
@@ -34,6 +34,7 @@ jobs:
             "*export*",
             "*decoder*",
             "*seq2seq*",
+            "*qwen3_asr*",
             "*funasr*",
             "*modeling*",
             "*diffusion*",
@@ -71,10 +72,15 @@ jobs:
         run: |
           uv pip install transformers==${{ matrix.transformers-version }}
 
-      - if: ${{ matrix.transformers-version == '4.57.6' }}
-        name: Install qwen-asr funasr packages
+      - if: ${{ matrix.transformers-version == '4.57.6' && matrix.test-pattern == '*qwen3_asr*' }}
+        name: Install qwen-asr package
         run: |
-          uv pip install qwen-asr funasr
+          uv pip install qwen-asr
+
+      - if: ${{ matrix.transformers-version == '4.57.6' && matrix.test-pattern == '*funasr*' }}
+        name: Install funasr package
+        run: |
+          uv pip install funasr
 
       - if: ${{ matrix.transformers-version == '4.45.0' }}
         name: Install specific dependencies and versions required for older transformers
